requirements.txt
altgraph==0.17.3
keyboard==0.13.5
mouse==0.7.1
pefile==2023.2.7
Pillow==10.0.0
playsound==1.2.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
pywin32==306
pywin32-ctypes==0.2.2
